LayoutTests/fast/scrolling/mac/scrollend-event-on-select-element-expected.txt


Test scroll over content
PASS overflowScrollendEventCount is 1
PASS windowScrollendEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
